Add support for fields to http_extractor.
The fields are a comma separated list of fields to extract. Special values
are:

body:N - Grab the first N bytes of the request and response body
uri - Grab the URI
status - Grab the response status

Everything else is treated as a header to look for.

Examples:

body:100,uri,status,host,content-length
x-forwarded-for,user-agent,server