Rewrite gh0st decoder.
The decoder does not have any hard-coded flags anymore. Instead of providing
a list of flags to search for we define a window size (default is 20) to
search in. We search the window for a dword that matches the packet size. If
we find a match we go back and grab the flag and parse it as if it were
a gh0st connection.

While the webcam, audio and screen spying are all recognized the data
associated with them is not parsed out. It would be nice to save a video
of the webcam and screen spying (or just stream it somewhere else). It
would also be nice to be able to save a copy of the captured audio.