More cleanups.
Remove dead code.

Cleanup logic in some functions instead of them being if/else. Just return
in the short if case.

The decompress() function is no longer. It was just a wrapper around
lznt1.dCompressBuf() with a bunch of dead code inside it. I killed the
decompress() function and made all calls to it just use
lznt1.dCompressBuf().