More cleanups.
Rearrange imports.

Cleanup struct parsing to be easier to read. The stuff in portlist() needs
testing. I know it didn't work before because it was doing this:

(pid, proclen) = unpack("<IB",data[:5])[0]

which is not valid syntax.

Also fix a bug if we can not find dirstart to politely say so and carry on.

Fix the struct parsing in getHeaders() to be easier on the eyes. Instead of
parsing values out of order and one at a time, parse them in order and at
once.