Creation of ChopConfig.
ChopConfig is a class which merges a default configuration with an optional
config file, then with all of the rest of the options and arguments provided to
things like chopshop and chopweb. The resulting object can be used for
configuring chopshop, chopweb, etc. directly.

I cleaned up chopshop now that the options are merged prior to configuring
choplib and chopui. Also added the config file options to chopweb.

The config file itself could use some organizing if we want to do so.