Introduce SSLim and ssl module.
The 'sslim' (pronounced slim) library is used to decrypt SSL sessions given
the private key. It only handles a few cipher suites and is intentionally
very minimal. It has only received very lightweight testing. If you want
information on how to use it I suggest checking out the docs at the top of
it.

The 'ssl' module is an example of using it. It will take in TCP and output
a 'raw' ChopProtocol object.