Add cert and key and 2 SSL pcaps.
Before anyone freaks out, this is a throw-away key I generated specifically
for these PCAPs. Nobody in their right mind should ever use this for anything
sensitive. I'm committing the key here because it is needed to decrypt the
example PCAPs also included here.

I'm including the cert because why not? ;)