Rename ssl.py to chop_ssl.py and handle raw.
If the filename is called 'ssl' it conflicts with something in M2Crypto.

Teach the http and payloads module to process 'raw' ChopProtocol objects.

This is mostly structural changes and because I'm OCD some whitespace
cleanup. I also simplied file extraction in http_extractor to be less
crazy to read.

Here are examples of how to use them:

./chopshop -f ssl-tests/ssl3.pcap "chop_ssl -k ssl-tests/server.key -v | payloads -t -u"
./chopshop -s . -f ssl-tests/ssl3.pcap "chop_ssl -k ssl-tests/server.key -v | http | http_extractor -s"