Commit initial version of etumbot.
This decoder can be used to decrypt traffic from etumbot. It is based upon the
Arbor report[1] and is only very lightly tested. There are lots of things left
still to do and they are all marked with 'XXX' in the code.

I'm looking for PCAPs that can be used to finish the missing features.